Create a list of parameters

The function creates a list of user-defined parameters.

Usage

parameters_deterministic_model(
  YL = 365,
  DT = 1,
  time_period = 365,
  NP = 21,
  agec = default_demog$agec,
  death = default_demog$death,
  season = FALSE,
  age_per = 1,
  N_human = 3e+07,
  incub = 5.5,
  inf_per = 6,
  nu = 1/21,
  Omega = 1,
  DeltaBase = 0.2,
  Sigma = 0.025,
  Epsilon = 1/19,
  Rm = 2.69,
  Mwt_mean = 1.5,
  Mwt_cv = 0.15,
  eip_mean = 8.4,
  Kappa = 0.5,
  Beta_hm_1 = 0.7,
  Beta_mh_1 = 0.7,
  propMwtControl = 0,
  TimeMwtControlOn = 1.5,
  TimeMwtControlOff = 2.5,
  Wb_cyto = 1,
  Wb_mat = 1,
  Wb_fM = 0.95,
  Wb_fF = 0.95,
  Wb_relsusc1 = 0.9,
  Wb_relinf1 = 0.5,
  Wb_starttime = 1,
  Wb_introlevel = 0,
  Wb_introduration = 60,
  vacc_child_coverage = 0,
  vacc_child_starttime = 30,
  vacc_child_stoptime = 30,
  vaccine_child_age = NULL,
  vacc_cu_coverage = 0,
  vacc_cu_time = 30,
  vaccine_cu_age = c(0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0),
  vacceff_prim = 0.75,
  rho_prim = 1,
  other_foi = 0.025,
  other_prop_immune = 0,
  propTransGlobal = 5e-04,
  propTransNN = 0,
  BG_FOI = 1e-08,
  phi_prim = 1,
  dis_pri = 0.2,
  mc_baseline = 2e-04,
  AGE_REC = 2,
  PropDiseaseReported = 1
)

Arguments

YL

Duration of a calendar year. Default = 365.

DT

Time step size. Default = 1.

time_period

Length of simulation. Default = 365.

NP

Number of patches. Default = 21.

agec

Numeric vector of age group widths. Default = c(1, 9, 10, 10, 10, 10, 10, 10, 10, 10, 10).

death

Numeric vector of mortality rates. Default = deathrt <- c(1e-10, 1e-10, 1e-10, 0.00277068683332695, 0.0210680857689784, 0.026724997685722, 0.0525354529367476, 0.0668013582441452, 0.119271483740379, 0.279105747097929, 0.390197266957464).

season

Logical for controlling the effect of seasonality. TRUE = maximum effect of seasonality. FALSE = no effect of seasonality. Default = FALSE

age_per

Time (weeks) between age updates. Default = 1.

N_human

Human population in each patch. Default = 30000000.

incub

Intrinsic Incubation Period (days). Default = 5.5.

inf_per

Total duration of human infectiousness (days). Default = 6.

nu

Inverse of virus generation time (1 / days). Virus generation time = serial interval. Default = 1 / 21.

Omega

Intensity of density dependence of mosquito larvae mortality rate. 0 = no density dependence. Default = 1.

DeltaBase

Adult mosquito mortality rate. Default = 0.2.

Sigma

Larval mosquito mortality rate. Default = 0.025

Epsilon

Larval mean development rate (1 / larvae mean development time in days). Default = 1/19.

Rm

Mosquito reproduction number (based on adult female fecundity and mortality rates). Default = 2.69.

Mwt_mean

Mean of the lognorm distribution of the at-equilibrium number of adult female mosquitos per person, per patch, without seasonality. Default = 1.5.

Mwt_cv

Standard deviation of the lognorm distribution of the at-equilibrium number of adult female mosquitos per person, per patch, without seasonality. Default = 0.15

eip_mean

Mean Extrinsic Incubation Period (days). Default = 8.4.

Kappa

Mosquito biting rate. Default = 0.5.

Beta_hm_1

Per bite transmission probability from humans to mosquitoes. Default = 0.7. This value is assigned to give a mean reproduction number, R0, across patches of 2.3 (with seasonal forcing).

Beta_mh_1

Per bite transmission probability from mosquitoes to humans. Arbitrarily assigned. Default = 0.7.

propMwtControl

Increase in mortality of adult wild type mosquitoes induced by a general type of intervention. Default = 0.

TimeMwtControlOn

Year of start of control of adult wild type mosquitoes. Default = 2.

TimeMwtControlOff

Year of end of control of adult wild type mosquitoes. Default = 3.

Wb_cyto

Degree of cytoplasmic incompatibility induced by Wolbachia. Default = 1.

Wb_mat

Degree of vertical transmission of Wolbachia. Default = 1.

Wb_fM

Increase in mortality induced by Wolbachia. Default = 0.95.

Wb_fF

Reduction in fecundity induced by Wolbachia. Default = 0.95.

Wb_relsusc1

Infectivity of a human host to Wolbachia infected mosquitoes (time T after host infection). Default = 0.9.

Wb_relinf1

Infectiousness of Wolbachia-infected mosquitoes time T after infection. Default = 0.75.

Wb_starttime

Time of first release of Wolbachia-infected mosquiotes (years). Default = 1.

Wb_introlevel

Ratio of Wolbachia-infected to wild type mosquitoes at introduction. (Not the proportion of Wolbachia AFTER introduction). Default = 0.

Wb_introduration

Duration of Wolbachia release (days). Default = 60.

vacc_child_coverage

Proportion of children vaccinated. Default = 0.

vacc_child_starttime

Time when vaccination starts. Default = 30.

vacc_child_stoptime

Time when vaccination ends. Default = 30.

vaccine_child_age

Vector of binary indicators for routine vaccination of age groups. 1 = vaccinate, 0 = do not vaccinate. Same length as agec. Default = NULL.

vacc_cu_coverage

Proportion of children who undergo catch up vaccination. Default = 0.

vacc_cu_time

Time when catch up vaccination occurs. Default = 30.

vaccine_cu_age

Vector of binary indicators for catch up vaccination of age groups. 1 = vaccinate, 0 = do not vaccinate. Same length as agec. Default = c(0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0), from 9 to 49.

vacceff_prim

Efficacy of vaccination in reducing infection. Default = 0.75.

rho_prim

Decay of protection from vaccine. Default = 1.

other_foi

Default = 0.025.

other_prop_immune

Propotion of population with pre-existing immunity. Default = 0.

propTransGlobal

Proportion of transmission between all patches. Default = 0.0005.

propTransNN

Proportion of transmission with nearest-neighbor patches. Default = 0.

BG_FOI

Force of infection on humans resulting from imported cases in travelers visiting from elsewhere. Default = 1e-8.

phi_prim

Degree of cross-immunity (not applicable for Zika). Default = 1.

dis_pri

Proportion of infections which are symptomatic. Default = 0.2.

mc_baseline

Baseline per birth microcephaly incidence rate. Default = 0.0002.

AGE_REC

Default = 2.

PropDiseaseReported

Reporting rate of symptomatic cases. Default = 0.1.

Value

Parameter list